+++ test/ext-test/gpu/cuda_ext.jl
@@ -5,16 +5,16 @@
     ψdi = Qobj(Int64[1, 0])
     ψdf = Qobj(Float64[1, 0])
     ψdc = Qobj(ComplexF64[1, 0])
-    ψsi = dense_to_sparse(ψdi)
-    ψsf = dense_to_sparse(ψdf)
-    ψsc = dense_to_sparse(ψdc)
+    ψsi = to_sparse(ψdi)
+    ψsf = to_sparse(ψdf)
+    ψsc = to_sparse(ψdc)
 
     Xdi = Qobj(Int64[0 1; 1 0])
     Xdf = Qobj(Float64[0 1; 1 0])
     Xdc = Qobj(ComplexF64[0 1; 1 0])
-    Xsi = dense_to_sparse(Xdi)
-    Xsf = dense_to_sparse(Xdf)
-    Xsc = dense_to_sparse(Xdc)
+    Xsi = to_sparse(Xdi)
+    Xsf = to_sparse(Xdf)
+    Xsc = to_sparse(Xdc)
 
     @test_throws DomainError cu(ψdi; word_size = 16)
 
@@ -67,19 +67,19 @@
     @test typeof(CuSparseMatrixCSR{ComplexF32}(Xsc).data) == CuSparseMatrixCSR{ComplexF32,Int32}
 
     # Sparse To Dense
-    # @test sparse_to_dense(cu(ψsi; word_size = 64)).data isa CuVector{Int64} # TODO: Fix this in CUDA.jl
-    @test sparse_to_dense(cu(ψsf; word_size = 64)).data isa CuVector{Float64}
-    @test sparse_to_dense(cu(ψsc; word_size = 64)).data isa CuVector{ComplexF64}
-    # @test sparse_to_dense(cu(Xsi; word_size = 64)).data isa CuMatrix{Int64} # TODO: Fix this in CUDA.jl
-    @test sparse_to_dense(cu(Xsf; word_size = 64)).data isa CuMatrix{Float64}
-    @test sparse_to_dense(cu(Xsc; word_size = 64)).data isa CuMatrix{ComplexF64}
+    # @test to_dense(cu(ψsi; word_size = 64)).data isa CuVector{Int64} # TODO: Fix this in CUDA.jl
+    @test to_dense(cu(ψsf; word_size = 64)).data isa CuVector{Float64}
+    @test to_dense(cu(ψsc; word_size = 64)).data isa CuVector{ComplexF64}
+    # @test to_dense(cu(Xsi; word_size = 64)).data isa CuMatrix{Int64} # TODO: Fix this in CUDA.jl
+    @test to_dense(cu(Xsf; word_size = 64)).data isa CuMatrix{Float64}
+    @test to_dense(cu(Xsc; word_size = 64)).data isa CuMatrix{ComplexF64}
 
-    # @test sparse_to_dense(Int32, cu(ψsf; word_size = 64)).data isa CuVector{Int32} # TODO: Fix this in CUDA.jl
-    # @test sparse_to_dense(Float32, cu(ψsf; word_size = 64)).data isa CuVector{Float32} # TODO: Fix this in CUDA.jl
-    # @test sparse_to_dense(ComplexF32, cu(ψsf; word_size = 64)).data isa CuVector{ComplexF32} # TODO: Fix this in CUDA.jl
-    # @test sparse_to_dense(Int64, cu(Xsf; word_size = 32)).data isa CuMatrix{Int64} # TODO: Fix this in CUDA.jl
-    # @test sparse_to_dense(Float64, cu(Xsf; word_size = 32)).data isa CuMatrix{Float64} # TODO: Fix this in CUDA.jl
-    # @test sparse_to_dense(ComplexF64, cu(Xsf; word_size = 32)).data isa CuMatrix{ComplexF64} # TODO: Fix this in CUDA.jl
+    # @test to_dense(Int32, cu(ψsf; word_size = 64)).data isa CuVector{Int32} # TODO: Fix this in CUDA.jl
+    # @test to_dense(Float32, cu(ψsf; word_size = 64)).data isa CuVector{Float32} # TODO: Fix this in CUDA.jl
+    # @test to_dense(ComplexF32, cu(ψsf; word_size = 64)).data isa CuVector{ComplexF32} # TODO: Fix this in CUDA.jl
+    # @test to_dense(Int64, cu(Xsf; word_size = 32)).data isa CuMatrix{Int64} # TODO: Fix this in CUDA.jl
+    # @test to_dense(Float64, cu(Xsf; word_size = 32)).data isa CuMatrix{Float64} # TODO: Fix this in CUDA.jl
+    # @test to_dense(ComplexF64, cu(Xsf; word_size = 32)).data isa CuMatrix{ComplexF64} # TODO: Fix this in CUDA.jl
 
     # brief example in README and documentation
     N = 20